rename: preadmissionprocessor dir -> preadmitter
